Overhaul of grammer to cover some whitespace handling edgecases.
This is the 2nd attempt to fix the whitespace handling.
The solution seems a little more robust, however the options rules are a
little more manual than I'd like. Had to really push the atomics to get
things like comments and linebreaks working throughout the map.

Tests are passing, other than the ignored ones (which depend on code yet
to be written). Should fix #1 (again).