Stubbed some of the internal er APIs.
Note: it is becoming apparent some of the early design choices will need
to be revised, mainly around how we handle `FormatOption`. It's lookling
like this enum will need a typed inner value for each variant, and there
will need to be a way to map each variant to the string name used in
graphviz and the er files.