Refactor types around the formatting options.
Seems like the option values should be an enum with inner values of an
appropriate type for each case. Part of what's tricky here is the
original Haskell used types in each case that aligned with the graphviz
lib they had, but since we're targeting dot lang as our output *and* the
dot lib we have expects us to prepare what boils down to html strings
using these as attributes, all these options are effectively "string" if
we want to align with html attributes...

I think we should type the options a little more strictly so that we can
provide better feedback to the caller before handing our dot lang over to the
dot cli.

So, for now, the intent will be to parse the raw strings from our ER
file, then cast them back to string for our html output. Okay.