Fixup integer option value parsing.
Originally, I was confused about those `Word8` values. I looked at the
haskell docs and saw that this was an unsigned int, but they also talked
about how this corresponds to a char (nothing out of the ordinary about
a u8 being a representation for a char).

Anyway, I guess that idea stuck in my head and for some reason I thought
we were meant to be reading only 1 char from the string value (which
seemed odd, but okay).

After reviewing the merge commit diff for the impls branch, I suddenly
realized that really we're just reading a string (multiple chars) and
parsing as int. That's all that's going on here.